Rename 'aux.rs' since it is an invalid filename on Windows
